move uri handling to own file
